pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "spacy>=3.8.5",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "docx>=0.2.4",
    "trafilatura>=2.0.0",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "flask-wtf>=1.2.2",
]
